Fixed order processing to properly handle changed order submissions
Order event email notifications would trigger a fatal error in cases where the order was submitted for processing where payment processing failed and the order was changed before being submitted again. In these cases, the original order balance and items need voided (voided order event) which triggered the email notifications.

Several fixes were put in place to correct the behaviors.

First, behaviors were added to correctly detect changes to the order by checking the state of items in the cart when the order is submitted and storing that state in the session to check against in the next order submission attempt.

Secondly, the Purchase Theme API context is setup early enough to handle the event notifications so that the event notifications can be disabled during order balance adjustments (voided event, invoiced event) to prevent superfluous emails from being sent to the customer.