Add API functions for finding discount amounts
Originally I wrote this for a client who needed a way to show  his
catalog discounts in the cart so that customers can see the whole
savings of Item Sale Prices + Cart Item Promotions + Catalog Promotions.

There are some changes that I made to the cart.php template file that
I'm not sure are necessary, but I will post the code online for review.