Added redirect to login function
This is a bad way of doing this I know, but I thought I would put in the request. Basically if a user has to request a password reminder during checkout, when they finally login there is no redirect to the checkout process. Can we therefore allow a developer to override this by adding a 'redirect' variable to the login form?